# src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Rating API
  version: 1.0.0
paths:
  /api/v1/ratings/{id}:
    patch:
      summary: Update a rating by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
            examples:
              example1:
                summary: Update Rating Example 1
                value:
                  name: "Updated Name 1"
                  comment: "Updated Comment 1"
                  points: 4
                  NA: "Updated NA 1"
              example2:
                summary: Update Rating Example 2
                value:
                  name: "Updated Name 2"
                  comment: "Updated Comment 2"
                  points: 5
                  NA: "Updated NA 2"
              example3:
                summary: Update Rating Example 3
                value:
                  name: "Updated Name 3"
                  comment: "Updated Comment 3"
                  points: 3
                  NA: "Updated NA 3"
      responses:
        '200':
          description: Rating updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        '400':
          description: Invalid JSON
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Invalid JSON"
        '404':
          description: Rating not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Rating not found"
components:
  schemas:
    Rating:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        comment:
          type: string
        points:
          type: integer
          maximum: 5
        NA:
          type: string
      example:
        id: 1
        name: "John Doe"
        comment: "This is a comment"
        points: 5
        NA: "Not Available"